working on new netzschleuder implementation
