+++ src/pathpyG/io/__init__.py
@@ -1,2 +1,3 @@
-from pathpyG.io.netzschleuder import read_netzschleuder_network, read_netzschleuder_record, read_graphtool, list_netzschleuder_records
+from pathpyG.io.netzschleuder import read_netzschleuder_graph, read_netzschleuder_record, list_netzschleuder_records
+from pathpyG.io.graphtool import read_graphtool
 from pathpyG.io.pandas import df_to_graph, df_to_temporal_graph, read_csv_graph, read_csv_temporal_graph, write_csv, add_node_attributes, add_edge_attributes, graph_to_df, temporal_graph_to_df
